move spores() and blossoms() to character.cpp
