+++ src/player.cpp
@@ -11260,28 +11260,6 @@ std::vector<std::string> player::get_overlay_ids() const
     return rval;
 }
 
-void Character::spores()
-{
-    fungal_effects fe( *g, g->m );
-    //~spore-release sound
-    sounds::sound( pos(), 10, sounds::sound_t::combat, _( "Pouf!" ), false, "misc", "puff" );
-    for( const tripoint &sporep : g->m.points_in_radius( pos(), 1 ) ) {
-        if( sporep == pos() ) {
-            continue;
-        }
-        fe.fungalize( sporep, this, 0.25 );
-    }
-}
-
-void Character::blossoms()
-{
-    // Player blossoms are shorter-ranged, but you can fire much more frequently if you like.
-    sounds::sound( pos(), 10, sounds::sound_t::combat, _( "Pouf!" ), false, "misc", "puff" );
-    for( const tripoint &tmp : g->m.points_in_radius( pos(), 2 ) ) {
-        g->m.add_field( tmp, fd_fungal_haze, rng( 1, 2 ) );
-    }
-}
-
 float player::power_rating() const
 {
     int dmg = std::max( { weapon.damage_melee( DT_BASH ),
